More changes before filtering and sorting
Before we can start performing the queries with filtering and sorting, some additional changes were made on the way. For one, choosing the categories to sort and filter by are ready. Next, returning to default settings is also possible. Some style was done to all unordered list and password fields now display that they're for passwords. Finally, registration was updated so that it will not register the customer i their username is too short.