Cart Update
The cart can now update the quantity of a product. Alongside that, displaying whether or not a customer has anything in their cart has been fixed. All that remains to the cart is the delete, validating the new quantity from the cart page and checking that there is enough stock after updating from both the cart and product page.